+++ src/middleware/middleware.js
@@ -4,9 +4,18 @@ import * as actions from '../actions/actions.js';
 function changeBrowserURL(action){
 
     const option = action.option;
+    if (option  === 'scroll'){
+        const path = action.router.path || '/';
+        const pos = utils.scrollpos[path] || 0;
+        setTimeout(()=>{
+            document.body.scrollTop = document.documentElement.scrollTop = pos;
+        },0);
+    } else {
+        document.body.scrollTop = document.documentElement.scrollTop = 0;
+    }
+
     switch (option) {
         case ('silent'):
-            document.body.scrollTop = document.documentElement.scrollTop = 0;
             history.replaceState(utils.navindex, null, action.router.previous);
             return;
         case ('popEvent'): //pop event already poped the url
@@ -14,11 +23,17 @@ function changeBrowserURL(action){
         default:
             utils.navindex++;
             history.pushState(utils.navindex, null, action.router.url);
-            document.body.scrollTop = document.documentElement.scrollTop = 0;
     }
 
 }
 
+
+function storeScroll (path){
+
+    path = path || '/';
+    utils.scrollpos[path] = document.body.scrollTop;
+
+}
 
 export function middleware ({ dispatch, getState }) {
     return (next) => {
@@ -26,6 +41,8 @@ export function middleware ({ dispatch, getState }) {
             //the main action concerning the user
             if (action.type === 'ROUTER_NAVIGATION'){
                 if (__CLIENT__) {
+
+                    storeScroll(getState().router.path);
                     if (history.pushState) { //fallback to Refresh
                         changeBrowserURL(action);
                     } else if (action.option !== 'popEvent') {